src/requirements.txt
flask
coverage
pdoc3
openpyxl
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
selenium==4.9.0
robotframework
robotframework-seleniumlibrary 
matplotlib
scipy